requirements.txt
fastapi
uvicorn
requests
python-dotenv
google-generativeai
httpx[http2]
openai